refactor: Changed folder structure according to the context of the files (#4)
