+++ classes/ManizalesWeather.py
